Role Backend_Correccion_2

+++ src/main/java/pe/edu/upc/connection2connection/repositories/IRoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.connection2connection.entities.Role;
 
 @Repository
-public interface IRoleRepository extends JpaRepository<Role, Long> {
+public interface IRoleRepository extends JpaRepository<Role, Integer> {
 
 }
